reformatted and added feature. at once
